fix weekDiagram logic
