+++ app/src/main/java/com/example/habitstracker/statistic/presentation/components/drawStatisticContainers.kt
@@ -1,0 +1,154 @@
+package com.example.habitstracker.statistic.presentation.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
+import com.example.habitstracker.habit.domain.DateHabitEntity
+import com.example.habitstracker.history.presentation.components.statistic_containers.CustomBlank
+import com.example.habitstracker.history.presentation.components.statistic_containers.getFilledBlankList
+import java.time.DayOfWeek
+import java.time.LocalDate
+import java.time.temporal.TemporalAdjusters
+
+@Composable
+fun DrawStatisticContainers(dateHabitList: List<DateHabitEntity>) {
+    val context = LocalContext.current
+
+    /** current streak and the best streak*/
+    val currentStreak = getCurrentStreak(dateHabitList)
+    val bestStreak = getBestStreak(dateHabitList)
+
+    /** total completed habits  */
+    val totalCompletedHabits = dateHabitList.count { it.isCompleted }
+
+
+    /** total completed habits this week */
+    val currentDate = LocalDate.now()
+    val lastMonday =
+        remember(currentDate) {
+            currentDate.with(TemporalAdjusters.previousOrSame(DayOfWeek.MONDAY))
+        }
+    val thisWeekSelectedHabits = dateHabitList
+        .filter { it.currentDate >= lastMonday.toString() }
+        .count { it.isCompleted }
+
+
+    /** Percentage of completed habits */
+    val totalHabits = dateHabitList.size
+    val percentage = totalCompletedHabits / totalHabits.toFloat() * 100
+
+    /** Perfect days */
+    val perfectDaysCounter by remember {
+        mutableIntStateOf(
+            dateHabitList
+                .groupBy { it.currentDate }
+                .count { (_, habits) -> habits.all { it.isCompleted } }
+        )
+    }
+
+    val thisWeekPerfectedDays by remember {
+        mutableIntStateOf(
+            dateHabitList
+                .filter { it.currentDate >= lastMonday.toString() }
+                .groupBy { it.currentDate }
+                .count { (_, habits) -> habits.all { it.isCompleted } }
+        )
+    }
+
+    LazyRow(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(vertical = 16.dp)
+            .wrapContentHeight(),
+        horizontalArrangement = Arrangement.Start,
+        verticalAlignment = Alignment.Top
+    ) {
+        val statsList =
+            getFilledBlankList(
+                context = context,
+                currentStreak = currentStreak,
+                bestStreak = bestStreak,
+                totalCompletedHabits = totalCompletedHabits,
+                thisWeekSelectedHabits = thisWeekSelectedHabits,
+                totalHabits = totalHabits,
+                percentage = percentage,
+                perfectDaysCounter = perfectDaysCounter,
+                thisWeekPerfectedDays = thisWeekPerfectedDays
+            )
+
+        statsList.forEach { blankItem ->
+            item {
+                CustomBlank(
+                    gradientColor = blankItem.gradientColor,
+                    topText = blankItem.topText,
+                    middleText = blankItem.middleText,
+                    bottomText = blankItem.bottomText,
+                )
+            }
+        }
+    }
+}
+
+private fun getBestStreak(streakList: List<DateHabitEntity>): Int {
+    if (streakList.isEmpty()) return 0
+
+    // Group all habit entries by their date
+    val mapDateToHabits = streakList.groupBy { it.currentDate }
+
+    // Sort dates descending (newest first)
+    val sortedDates = mapDateToHabits.keys.sorted()
+
+    var bestStreak = 0
+    var currentStreak = 0
+    var previousDate: LocalDate? = null
+
+    for (date in sortedDates) {
+        val habitsForDate = mapDateToHabits[date].orEmpty()
+        val total = habitsForDate.size
+        val completed = habitsForDate.count { it.isCompleted }
+
+        val allCompleted = (total == completed)
+
+        if (allCompleted) {
+            if (previousDate == null || previousDate.plusDays(1) == LocalDate.parse(date)) {
+                currentStreak++
+            } else {
+                currentStreak = 1
+            }
+            bestStreak = maxOf(bestStreak, currentStreak)
+        } else {
+            currentStreak = 0
+        }
+
+        previousDate = LocalDate.parse(date)
+    }
+
+    return bestStreak
+}
+
+private fun getCurrentStreak(streakList: List<DateHabitEntity>): Int {
+    if (streakList.isEmpty()) return 0
+
+    var streak = 0
+    val groupedStreakList = streakList.groupBy { it.currentDate }
+
+    groupedStreakList.forEach { mapDateAndHabits ->
+        val habitsCount = mapDateAndHabits.value.count()
+        val isCompletedCount = mapDateAndHabits.value.count { it.isCompleted }
+        if (habitsCount == isCompletedCount)
+            streak++
+        else return streak
+    }
+
+    return streak
+}
